Fix many TODOs
adl-frontend : 
 - Add various specific errors
 - Remove "expectedKind" in DefinitionReference AST nodes
 - Improve checking in ADL JTBProcessor : raise an error in case of multiple "data" specifications and in case of a sub-component that references a template parameter and has an anonymous definition.

adl-backend : Fix support for 'executable-name'. The exec-name is not required to be a valid absolute path (as defined in PathHelper)

common-frontend : Fix support for escaped characters in evaluation of StringLitteral

cpl-preproc : CPLChecker throw ADLException with specific errors. The grammar adds the error locator depending on the error template returned by CPLChecker.


git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/trunk@645 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0