OPEN - issue MIND-60: '.h' files included in IDL are not found by GCC http://jira.ow2.org/browse/MIND-60
idl-frontend : Fix bug in IncludeHeaderResolver. try to resolve included .h files relatively to encapsulating directory. If found, translate include path to absolute path; otherwise leave as it is and suppose that it will be resolved by GCC (thanks to a '-I' c-flag)

idl-backend : Update tests to check this issue.

adl-backend : update helloworld.Service to include a .h file and test this issue.

git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/trunk@803 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0