refactor functional tests of adl-backend :
Do not use maven to unpack dependencies (it causes various problems). The dependencies (fractal-runtime) are unpacked directly by the test framework (in target/test-deps). To do so, a resource that is supposed to be in the dependency JAR is researched in the classpath ("fractal/api/Component.itf" for instance) and the JAR that contains it, is unpacked (if the resource is not in a JAR then the directory that contains it is used).

This also ease the configuration of modules that reuse adl-backend tests (cf. annotations package).

git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/trunk@993 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0