IN PROGRESS - issue MIND-51: Improve error handling in mindc http://jira.ow2.org/browse/MIND-51
Update compilation framework. The exec method of a compilation command returns a boolean indicating if the command has been executed correctly. The CompilationCommandExecutor use this boolean to determine if the execution of subsequent commands must be performed. It handle two policies : a fail-fast policy where the compilation is stopped at the first failure; and a default policy where compilation continue.

Add semantic error tests in adl-frontend

git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/branches/MIND-51@1090 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0