Various bug fixes and improvements
In mindc assembly: add a "ext" folder in binary assembly where compiler extensions can be put. This will allow to separate the JARs that belongs to the core compiler to the JARs that are provided by third parties. The 'mindc' and 'mindc.bat' have been updated to add JARs in the "ext" folder to the classpath.

git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/branches/guice@1377 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0