Add "attribute-like" syntax for the declaration of private data:   - data foo.h:myStruct myDataField;
These declarations can be used in addition to former data declaration. It can be used to add private data fields to an inherited primitive definition.
If, in a given component, the "PRIVATE structure" (i.e. the former data declaration) contains a field with the same name as a data field declared in the ADL, then the latter is used and mask the one declared in PRIVATE structure.
If a component contains only data field declared in the ADL (no PRIVATE structure), then the Mind preprocessor may raise an error if the implementation code accesses a PRIVATE field that is not declared in the ADL with a clear error message.

git-svn-id: svn://svn.forge.objectweb.org/svnroot/mind/mindc/branches/mindc-2.0@1485 a01ce9dd-6950-4f94-9599-cd5ca46ef4f0